implementacija funkcionalnosti

+++ src/main/java/hr/adnanb/rbazadatak/service/impl/ClientServiceImpl.java
@@ -3,6 +3,8 @@ package hr.adnanb.rbazadatak.service.impl;
 import hr.adnanb.rbazadatak.domain.entity.Client;
 import hr.adnanb.rbazadatak.repository.ClientRepository;
 import hr.adnanb.rbazadatak.service.ClientService;
+import hr.adnanb.rbazadatak.service.KafkaProducerService;
+import jakarta.transaction.Transactional;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -16,6 +18,9 @@ public class ClientServiceImpl implements ClientService {
     @Autowired
     private ClientRepository clientRepository;
 
+    @Autowired
+    private KafkaProducerService kafkaProducerService;
+
     @Override
     public Client findByOib(String oib){
         return clientRepository.findByOib(oib);
@@ -27,6 +32,7 @@ public class ClientServiceImpl implements ClientService {
     }
 
     @Override
+    @Transactional
     public Client save(Client client){
          client = clientRepository.save(client);
          log.info("Client sucessfuly saved: {}", client);
@@ -34,8 +40,20 @@ public class ClientServiceImpl implements ClientService {
     }
 
     @Override
+    @Transactional
     public void deleteByOib(String oib){
+        Client client = this.findByOib(oib);
+        if(client == null){
+            throw new RuntimeException(new StringBuilder().append("Client with oib ")
+                    .append(oib).append(" does not exist!").toString());
+        }
+
         clientRepository.deleteByOib(oib);
+    }
+
+    @Override
+    public void sendClientToKafka(Client client){
+        kafkaProducerService.sendMessage("rbazadatak-topic", client.toString());
     }
 
 }
